requirements.txt
Flask>=2.0.0,<3.0.0
Werkzeug==2.0.3
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
pymongo==4.3.3
dnspython==2.3.0
bcrypt==4.0.1
certifi==2023.7.22
pymongo
# For AI integration (when you're ready)
# openai==0.27.0